web: improvements in the storage hooks
- Helpers do not modify the given apiModel object.
